Refactor repo and command pkgs in go scripts
